requirements_python.txt
asgiref==3.7.2
black==23.12.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
Django==5.0
django-filter==23.5
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
everapi==0.1.1
freecurrencyapi==0.1.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
Markdown==3.5.1
MarkupSafe==2.1.5
mypy-extensions==1.0.0
mysqlclient==2.2.1
openapi-codec==1.3.2
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.2
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.2.1
